Create new stepcount DB row only if necessary
New DB rows only need to be created if there are additional steps or
the previous interval must not be extended. The previous interval must
not be extended if it has > 0 steps, because then these steps would be
shifted to the current point in time.

This change is not necessary for correct behavior, but keeps the
database sparse.